+++ shared-components/frontend/package/design-system/ds/button/src/lib/button.stories.ts
@@ -4,7 +4,7 @@ import { ButtonComponent } from './button.component';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories
 const meta: Meta<ButtonComponent> = {
-  title: 'Example/Button',
+  title: 'Design System/Button',
   component: ButtonComponent,
   tags: ['autodocs'],
   argTypes: {