+++ shared-components/frontend/package/design-system/ds/icon/src/public-api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of icon
+ */
+
+export * from './lib/icon';
+export * from './lib/icon.registry';